% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/break_it.R
\name{break_it}
\alias{break_it}
\title{break it}
\usage{
break_it(name, break_it)
}
\arguments{
\item{name}{name of yourself}

\item{break_it}{do you wanna break it?}
}
\value{
a string if youre Smart
}
\description{
break it
}
