% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{go}
\alias{go}
\title{Lets go?}
\usage{
go(name, go)
}
\arguments{
\item{name}{your name for the lil guy}

\item{go}{boolean if ya wanna go}
}
\value{
statement declaring if were going
}
\description{
Lets go?
}
