% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statement.R
\name{statement}
\alias{statement}
\title{A function to state what you had consumed tonight. but with a twist... heheh... you can LIE!}
\usage{
statement(consumable, amount, lie)
}
\arguments{
\item{consumable}{name of whatever you had just consumed}

\item{amount}{total number of whatever you had just consumed}

\item{lie}{boolean to choose if you lie or not... heheheh..}
}
\description{
A function to state what you had consumed tonight. but with a twist... heheh... you can LIE!
}
\examples{
statement("apple", 3, TRUE)
}
